Move evidence tests to their own section.
